Rename Counter import types paths
